Experimenting with build process

+++ build.bat
@@ -1,13 +1,35 @@
 @ECHO OFF
-if exist build\ (
-    del /S /Q build\\*
-    rmdir /S /Q build
+
+set "release_type=%1"
+
+if "%release_type%"=="--debug" (
+    if exist build\ (
+        del /S /Q build\\*
+        rmdir /S /Q build
+    )
+    if exist dist\ (
+        del /S /Q dist\\*
+        rmdir /S /Q dist
+    )
+
+    echo debug> release_type.txt
+    python -m PyInstaller --noconfirm --clean build.spec
+    del release_type.txt
+) else (
+    if "%release_type%"=="--release" (
+        if exist build\ (
+            del /S /Q build\\*
+            rmdir /S /Q build
+        )
+        if exist dist\ (
+            del /S /Q dist\\*
+            rmdir /S /Q dist
+        )
+
+        echo release> release_type.txt
+        python -m PyInstaller --noconfirm --clean build.spec
+        del release_type.txt
+    ) else (
+        echo Invalid release type
+    )
 )
-if exist dist\ (
-    del /S /Q dist\\*
-    rmdir /S /Q dist
-)
-pyinstaller --name=MC3DS-Texture-Maker --onefile --hide-console hide-early --icon icon.ico --add-data assets:assets --add-data icon.ico:. --add-data icon.png:. --add-data icon2.png:. --hidden-import PIL --hidden-import PIL._imagingtk --hidden-import PIL._tkinter_finder --hidden-import pkg_resources.extern main.py
-pyinstaller --name=3DST-Viewer --onefile --windowed --icon icon.ico --hidden-import PIL --hidden-import PIL._imagingtk --hidden-import PIL._tkinter_finder --hidden-import pkg_resources.extern py3dstViewer.py
-del MC3DS-Texture-Maker.spec
-del 3DST-Viewer.spec
